edi: sort out components mixins, add input mixins
