+++ edi_oca/components/__init__.py
@@ -1,1 +1,3 @@
 from . import base
+from . import base_output
+from . import base_input
